// src/generated/requests/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type Product = {
    id: number;
    title: string;
    price: number;
    description: string;
    category: string;
    image: string;
};

export type GetAllProductsResponse = Array<Product>;

export type GetProductByIdData = {
    id: number;
};

export type GetProductByIdResponse = Product;

export type $OpenApiTs = {
    '/products': {
        get: {
            res: {
                /**
                 * OK
                 */
                200: Array<Product>;
            };
        };
    };
    '/products/{id}': {
        get: {
            req: GetProductByIdData;
            res: {
                /**
                 * OK
                 */
                200: Product;
            };
        };
    };
};